Initial commit of the code editor. Editor currently has syntax highlighting for c++, matching bracket highlighting, and line numbers.
A file can be loaded into the code editor from a local directory via the load icon. Saving is not implemented yet.

The editor can currently undo/redo, jump to a line, format text, and find/replace.

If you choose to load a file, it is loaded into the code editor, where you can then submit it.